requirements.txt
Flask
Flask-SocketIO
python-nmap
Flask-Login
Flask-SQLAlchemy
eventlet
